+++ src/main/java/lotto/LottoApplication.java
@@ -3,7 +3,6 @@ package lotto;
 import lotto.controller.LottoController;
 
 public class LottoApplication {
-
     public static void main(String[] args) {
         LottoController.run();
     }
