+++ src/main/java/lotto/domain/LottoPrize.java
@@ -1,4 +1,4 @@
-package lotto_step3.domain;
+package lotto.domain;
 
 import java.util.Arrays;
