--- src/main/java/lotto_step3/LottoApplication.java
@@ -1,9 +1,0 @@
-package lotto_step3;
-
-import lotto_step3.controller.LottoController;
-
-public class LottoApplication {
-    public static void main(String[] args) {
-        LottoController.run();
-    }
-}
