+++ src/test/java/lotto/domain/LottosTest.java
@@ -1,4 +1,4 @@
-package lotto_step3.domain;
+package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
